refactor: WIP on selection saving as a place to roll back
